# openapi.yml
info:
  version: 1.0.0
  title: Agent Platform
  description: This is a specification for openapi using
  contact:
    name: Agent Platform UH Team
    email: agentPlatform@grs.uh.cu
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
paths:
  '/getSimilarAgents/{Name}':
    get:
      description: Get the agents that are similars to the agent passed as paramerter
      operationId: getSimilarAgent
      parameters:
        - in: path
          name: Name
          required: true
          description: Name of the Agent
          type: string
      responses:
        '200':
          description: Get the agent that are similar
          schema:
            $ref: '#/definitions/Agent'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/getAgentsForFunction/{Name}':
    get:
      description: Get the agents that match with the function name passed as params
      operationId: getAgentsByFunction
      parameters:
        - in: path
          name: Name
          required: true
          description: Name of the Function
          type: string
      responses:
        '200':
          description: >
            Get the agents endpoints([3]Addr) that represent the active agents
            that

            follow the function name
          schema:
            type: array
            items:
              type: array
              items:
                $ref: '#/definitions/Addr'
              maxItems: 3
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
  /getAllAgentsNames:
    get:
      description: |
        Get all agents names registered in the platforms
      operationId: getAgentsNames
      responses:
        '200':
          description: get all Agents
          schema:
            type: array
            items:
              type: string
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
      parameters: []
  /getPeers:
    get:
      description: |
        Return all peers connected to the platform network
      operationId: getPeers
      responses:
        '200':
          description: get Peers response
          schema:
            type: array
            items:
              $ref: '#/definitions/Addr'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
      parameters: []
  /registerAgent:
    post:
      description: Register a new Agent in the platform
      operationId: registerAgent
      responses:
        '200':
          description: Agent registered
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
      parameters:
        - description: Agent to register
          required: true
          name: body
          in: body
          schema:
            $ref: '#/definitions/Agent'
  '/getAgent/{Name}':
    get:
      description: Get the agent that follow a simple criteria
      operationId: getAgent
      parameters:
        - in: path
          name: Name
          required: true
          description: Name of the Agent
          type: string
      responses:
        '200':
          description: |
            Response Should contain 3 Addr
            Response[0] Agent Addr
            Response[1] Agent Is Alive endpoint Addr
            Response[2] Agent Documentation Addr
          schema:
            type: array
            items:
              $ref: '#/definitions/Addr'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'
swagger: '2.0'
schemes:
  - http
host: agentplatform.grs.uh.cu
basePath: /api/v1
definitions:
  Addr:
    type: object
    required:
      - Ip
      - Port
    properties:
      Ip:
        type: string
      Port:
        type: integer
        format: int64
  Agent:
    type: object
    required:
      - name
    properties:
      Name:
        type: string
      Function:
        type: string
      EndpointService:
        type: array
        items:
          $ref: '#/definitions/Addr'
      IsAliveService:
        type: object
        additionalProperties:
          $ref: '#/definitions/Addr'
      Documentation:
        type: object
        additionalProperties:
          $ref: '#/definitions/Addr'
      TestCases:
        type: array
        items:
          $ref: '#/definitions/TestCase'
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
  TestCase:
    type: object
    required:
      - Input
      - Output
    properties:
      Input:
        type: string
      Output:
        type: string
x-components: {}